bioinsight_ai/requirements.txt
aioboto3==13.4.0
aiobotocore==2.18.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.8.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
asyncer==0.0.7
attrs==25.1.0
azure-core==1.32.0
azure-storage-blob==12.24.1
azure-storage-file-datalake==12.18.1
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bidict==0.23.1
bleach==6.2.0
boto3==1.36.1
botocore==1.36.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chainlit>=2.5.5
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dicom2nifti==2.6.1
dirtyjson==1.0.8
distro==1.9.0
docutils==0.21.2
duckdb==1.2.1
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.1
frozenlist==1.5.0
fsspec==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idc-index==0.9.0
idc-index-data==21.0.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
inflection==0.5.1
ipydatawidgets==4.3.2
ipykernel==6.29.5
ipympl==0.9.7
ipython==8.21.0
ipython-genutils==0.2.0
ipywidgets==7.8.5
isodate==0.7.2
itk-core==5.4.3
itk-filtering==5.4.3
itk-meshtopolydata==0.11.1
itk-numerics==5.4.3
itkwidgets==0.32.6
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==1.1.11
kiwisolver==1.4.8
Lazify==0.4.0
lifelines==0.27.4
literalai==0.1.201
llama-cloud==0.1.13
llama-cloud-services==0.6.3
llama-index==0.12.41
llama-index-core==0.12.41
llama-index-cli==0.4.3
llama-index-agent-openai==0.4.9
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.4.0
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-bedrock==0.3.8
llama-index-llms-bedrock-converse==0.7.2
llama-index-retrievers-bedrock==0.3.0
llama-parse==0.6.2
logger==1.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mcp==1.9.2
mistune==3.1.3
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.29.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nltk==3.9.1
notebook==6.4.13
openai>=1.81.0,<2.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-alephalpha==0.40.7
opentelemetry-instrumentation-anthropic==0.40.7
#opentelemetry-instrumentation-bedrock==0.40.7
opentelemetry-instrumentation-chromadb==0.40.7
opentelemetry-instrumentation-cohere==0.40.7
opentelemetry-instrumentation-crewai==0.40.7
opentelemetry-instrumentation-google-generativeai==0.40.7
opentelemetry-instrumentation-groq==0.40.7
opentelemetry-instrumentation-haystack==0.40.7
opentelemetry-instrumentation-lancedb==0.40.7
opentelemetry-instrumentation-langchain==0.40.7
opentelemetry-instrumentation-llamaindex==0.40.7
opentelemetry-instrumentation-logging==0.50b0
opentelemetry-instrumentation-marqo==0.40.7
opentelemetry-instrumentation-mcp==0.40.7
opentelemetry-instrumentation-milvus==0.40.7
opentelemetry-instrumentation-mistralai==0.40.7
opentelemetry-instrumentation-ollama==0.40.7
opentelemetry-instrumentation-openai==0.40.7
opentelemetry-instrumentation-pinecone==0.40.7
opentelemetry-instrumentation-qdrant==0.40.7
opentelemetry-instrumentation-replicate==0.40.7
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-sagemaker==0.40.7
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-threading==0.50b0
opentelemetry-instrumentation-together==0.40.7
opentelemetry-instrumentation-transformers==0.40.7
opentelemetry-instrumentation-urllib3==0.50b0
opentelemetry-instrumentation-vertexai==0.40.7
opentelemetry-instrumentation-watsonx==0.40.7
opentelemetry-instrumentation-weaviate==0.40.7
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.50b0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.0
posthog==3.25.0
ppx==1.4.4
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydicom==3.0.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.3.1
pyradiomics==3.0.1
pyteomics==4.7.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-gdcm==3.0.25
python-multipart==0.0.18
python-Levenshtein==0.27.1
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
roman-numerals-py==3.1.0
rpds-py==0.25.1
rsa==4.9
s3transfer==0.11.3
s5cmd==0.2.0
Send2Trash==1.8.3
setuptools==80.9.0
simple-websocket==1.1.0
simpleitk==2.5.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.2.3
sphinx-click==6.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.41.3
striprtf==0.0.26
sympy==1.14.0
syncer==2.0.3
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tomli==2.2.1
torch==2.2.2
tornado==6.5.1
TotalSegmentator==2.9.0
tqdm==4.67.1
traceloop-sdk==0.40.7
traitlets==5.6.0
traittypes==0.2.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.29.0
urllib3==2.3.0
uvicorn==0.34.0
versioneer==0.29
watchfiles==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==3.6.10
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0